feat: Reorganize NMT repo (#319)
