+++ primitives/src/merkle_tree/namespaced_merkle_tree/mod.rs
@@ -5,23 +5,26 @@
 // along with the Jellyfish library. If not, see <https://mit-license.org/>.
 
 //! Implementation of a Namespaced Merkle Tree.
-use alloc::{
-    collections::{btree_map::Entry, BTreeMap},
-    vec,
-};
+use alloc::collections::{btree_map::Entry, BTreeMap};
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
-use ark_std::{string::ToString, vec::Vec};
+use ark_std::vec::Vec;
 use core::{borrow::Borrow, fmt::Debug, hash::Hash, marker::PhantomData, ops::Range};
-use itertools::Itertools;
-use serde::{Deserialize, Serialize};
 use typenum::Unsigned;
 
 use crate::errors::{PrimitivesError, VerificationResult};
+
+use self::{
+    hash::{NamespacedHash, NamespacedHasher},
+    proof::{NaiveNamespaceProof, NamespaceProofType},
+};
 
 use super::{
     append_only::MerkleTree, internal::MerkleProof, AppendableMerkleTreeScheme, DigestAlgorithm,
     Element, Index, LookupResult, MerkleCommitment, MerkleTreeScheme, NodeValue,
 };
+
+mod hash;
+mod proof;
 
 /// Namespaced Merkle Tree where leaves are sorted by a namespace identifier.
 /// The data structure supports namespace inclusion proofs.
@@ -45,25 +48,6 @@ where
         proof: &Self::NamespaceProof,
         namespace: Self::NamespaceId,
     ) -> Result<VerificationResult, PrimitivesError>;
-}
-
-/// NamespacedHasher wraps a standard hash function (implementer of
-/// DigestAlgorithm), turning it into a hash function that tags internal nodes
-/// with namespace ranges.
-#[derive(Debug, Clone)]
-pub struct NamespacedHasher<H, E, I, T, N>
-where
-    H: DigestAlgorithm<E, I, T>,
-    E: Element + Namespaced<Namespace = N>,
-    N: Namespace,
-    I: Index,
-    T: NodeValue,
-{
-    phantom1: PhantomData<H>,
-    phantom2: PhantomData<E>,
-    phantom3: PhantomData<I>,
-    phantom4: PhantomData<T>,
-    phantom5: PhantomData<N>,
 }
 
 /// Completeness proof for a namespace
@@ -126,92 +110,6 @@ impl Namespace for u64 {
     }
 }
 
-#[derive(
-    CanonicalSerialize,
-    CanonicalDeserialize,
-    Hash,
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    Ord,
-    Eq,
-    PartialEq,
-    PartialOrd,
-)]
-/// Represents a namespaced internal tree node
-pub struct NamespacedHash<T, N>
-where
-    N: Namespace,
-    T: NodeValue,
-{
-    min_namespace: N,
-    max_namespace: N,
-    hash: T,
-}
-
-impl<T, N> NamespacedHash<T, N>
-where
-    N: Namespace,
-    T: NodeValue,
-{
-    /// Constructs a new NamespacedHash
-    pub fn new(min_namespace: N, max_namespace: N, hash: T) -> Self {
-        Self {
-            min_namespace,
-            max_namespace,
-            hash,
-        }
-    }
-}
-
-impl<E, H, T, I, N> DigestAlgorithm<E, I, NamespacedHash<T, N>> for NamespacedHasher<H, E, I, T, N>
-where
-    E: Element + Namespaced<Namespace = N>,
-    I: Index,
-    N: Namespace,
-    T: NodeValue,
-    H: DigestAlgorithm<E, I, T> + BindNamespace<E, I, T, N>,
-{
-    // Assumes that data is sorted by namespace, will be enforced by "append"
-    fn digest(data: &[NamespacedHash<T, N>]) -> Result<NamespacedHash<T, N>, PrimitivesError> {
-        if data.is_empty() {
-            return Ok(NamespacedHash::default());
-        }
-        let first_node = data[0];
-        let min_namespace = first_node.min_namespace;
-        let mut max_namespace = first_node.max_namespace;
-        let mut nodes = vec![H::generate_namespaced_commitment(first_node)];
-        for node in &data[1..] {
-            if node == &NamespacedHash::default() {
-                continue;
-            }
-            // Ensure that namespaced nodes are sorted
-            if node.min_namespace < max_namespace {
-                return Err(PrimitivesError::InternalError(
-                    "Namespace Merkle tree leaves are out of order".to_string(),
-                ));
-            }
-            max_namespace = node.max_namespace;
-            nodes.push(H::generate_namespaced_commitment(*node));
-        }
-
-        let inner_hash = H::digest(&nodes)?;
-
-        Ok(NamespacedHash::new(
-            min_namespace,
-            max_namespace,
-            inner_hash,
-        ))
-    }
-
-    fn digest_leaf(pos: &I, elem: &E) -> Result<NamespacedHash<T, N>, PrimitivesError> {
-        let namespace = elem.get_namespace();
-        let hash = H::digest_leaf(pos, elem)?;
-        Ok(NamespacedHash::new(namespace, namespace, hash))
-    }
-}
-
 type InnerTree<E, H, T, N, Arity> =
     MerkleTree<E, NamespacedHasher<H, E, u64, T, N>, u64, Arity, NamespacedHash<T, N>>;
 
@@ -248,7 +146,7 @@ where
 
     fn from_elems(
         height: usize,
-        elems: impl IntoIterator<Item = impl core::borrow::Borrow<Self::Element>>,
+        elems: impl IntoIterator<Item = impl Borrow<Self::Element>>,
     ) -> Result<Self, PrimitivesError> {
         let mut namespace_ranges: BTreeMap<N, Range<u64>> = BTreeMap::new();
         let mut max_namespace = <N as Namespace>::min();
@@ -333,260 +231,6 @@ where
     ) -> Result<(), PrimitivesError> {
         // TODO(#314): update namespace metadata
         self.inner.push(elem)
-    }
-}
-
-/// Indicates whether the namespace proof represents a populated set or an empty
-/// set
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-enum NamespaceProofType {
-    Presence,
-    Absence,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(bound = "E: CanonicalSerialize + CanonicalDeserialize,
-                 T: CanonicalSerialize + CanonicalDeserialize,")]
-/// Namespace Proof
-pub struct NaiveNamespaceProof<E, T, Arity, N, H>
-where
-    E: Element + Namespaced<Namespace = N>,
-    T: NodeValue,
-    H: DigestAlgorithm<E, u64, T> + BindNamespace<E, u64, T, N>,
-    N: Namespace,
-    Arity: Unsigned,
-{
-    proof_type: NamespaceProofType,
-    // TODO(#140) Switch to a batch proof
-    proofs: Vec<MerkleProof<E, u64, NamespacedHash<T, N>, Arity>>,
-    left_boundary_proof: Option<MerkleProof<E, u64, NamespacedHash<T, N>, Arity>>,
-    right_boundary_proof: Option<MerkleProof<E, u64, NamespacedHash<T, N>, Arity>>,
-    first_index: u64,
-    phantom: PhantomData<H>,
-}
-impl<E, T, Arity, N, H> NamespaceProof for NaiveNamespaceProof<E, T, Arity, N, H>
-where
-    E: Element + Namespaced<Namespace = N>,
-    T: NodeValue,
-    H: DigestAlgorithm<E, u64, T> + BindNamespace<E, u64, T, N>,
-    N: Namespace,
-    Arity: Unsigned,
-{
-    type Leaf = E;
-    type Node = T;
-    type Namespace = N;
-
-    fn get_namespace_leaves(&self) -> Vec<&Self::Leaf> {
-        let num_leaves = match self.proof_type {
-            NamespaceProofType::Presence => self.proofs.len(),
-            NamespaceProofType::Absence => 0,
-        };
-        self.proofs
-            .iter()
-            // This unwrap is safe assuming that the proof is valid
-            .map(|proof| proof.elem().unwrap())
-            .take(num_leaves)
-            .collect_vec()
-    }
-
-    fn verify(
-        &self,
-        root: &NamespacedHash<T, N>,
-        namespace: N,
-    ) -> Result<VerificationResult, PrimitivesError> {
-        match self.proof_type {
-            NamespaceProofType::Presence => self.verify_presence_proof(root, namespace),
-            NamespaceProofType::Absence => self.verify_absence_proof(root, namespace),
-        }
-    }
-}
-
-impl<E, T, Arity, N, H> NaiveNamespaceProof<E, T, Arity, N, H>
-where
-    E: Element + Namespaced<Namespace = N>,
-    T: NodeValue,
-    H: DigestAlgorithm<E, u64, T> + BindNamespace<E, u64, T, N>,
-    N: Namespace,
-    Arity: Unsigned,
-{
-    fn verify_left_namespace_boundary(
-        &self,
-        root: &NamespacedHash<T, N>,
-        namespace: N,
-    ) -> Result<VerificationResult, PrimitivesError> {
-        if let Some(boundary_proof) = self.left_boundary_proof.as_ref() {
-            // If there is a leaf to the left of the namespace range, check that it is less
-            // than the target namespace
-            if boundary_proof
-                .elem()
-                .ok_or(PrimitivesError::InconsistentStructureError(
-                    "Boundary proof does not contain an element".into(),
-                ))?
-                .get_namespace()
-                >= namespace
-                || *boundary_proof.index() != self.first_index - 1
-            {
-                return Ok(Err(()));
-            }
-            // Verify the boundary proof
-            if <InnerTree<E, H, T, N, Arity>>::verify(root, boundary_proof.index(), boundary_proof)?
-                .is_err()
-            {
-                return Ok(Err(()));
-            }
-        } else {
-            // If there is no left boundary, ensure that target namespace is the tree's
-            // minimum namespace
-            if root.min_namespace != namespace {
-                return Ok(Err(()));
-            }
-        }
-        Ok(Ok(()))
-    }
-
-    fn verify_right_namespace_boundary(
-        &self,
-        root: &NamespacedHash<T, N>,
-        namespace: N,
-    ) -> Result<VerificationResult, PrimitivesError> {
-        if let Some(boundary_proof) = self.right_boundary_proof.as_ref() {
-            // If there is a leaf to the left of the namespace range, check that it is less
-            // than the target namespace
-            if boundary_proof
-                .elem()
-                .ok_or(PrimitivesError::InconsistentStructureError(
-                    "Boundary proof does not contain an element".to_string(),
-                ))?
-                .get_namespace()
-                <= namespace
-                || *boundary_proof.index() != self.first_index + self.proofs.len() as u64
-            {
-                return Ok(Err(()));
-            }
-            // Verify the boundary proof
-            if <InnerTree<E, H, T, N, Arity>>::verify(root, boundary_proof.index(), boundary_proof)?
-                .is_err()
-            {
-                return Ok(Err(()));
-            }
-        } else {
-            // If there is no left boundary, ensure that target namespace is the tree's
-            // minimum namespace
-            if root.max_namespace != namespace {
-                return Ok(Err(()));
-            }
-        }
-        Ok(Ok(()))
-    }
-
-    fn verify_absence_proof(
-        &self,
-        root: &NamespacedHash<T, N>,
-        namespace: N,
-    ) -> Result<VerificationResult, PrimitivesError> {
-        if namespace < root.min_namespace || namespace > root.max_namespace {
-            // Easy case where the namespace isn't covered by the range of the tree root
-            return Ok(Ok(()));
-        } else {
-            // Harder case: Find an element whose namespace is greater than our
-            // target and show that the namespace to the left is less than our
-            // target
-            let left_proof = &self.left_boundary_proof.as_ref().cloned().ok_or(
-                PrimitivesError::InconsistentStructureError(
-                    "Left Boundary proof must be present".into(),
-                ),
-            )?;
-            let right_proof = &self.right_boundary_proof.as_ref().cloned().ok_or(
-                PrimitivesError::InconsistentStructureError(
-                    "Right boundary proof must be present".into(),
-                ),
-            )?;
-            let left_index = left_proof.index();
-            let left_ns = left_proof
-                .elem()
-                .ok_or(PrimitivesError::InconsistentStructureError(
-                    "The left boundary proof is missing an element".into(),
-                ))?
-                .get_namespace();
-            let right_index = right_proof.index();
-            let right_ns = right_proof
-                .elem()
-                .ok_or(PrimitivesError::InconsistentStructureError(
-                    "The left boundary proof is missing an element".into(),
-                ))?
-                .get_namespace();
-            // Ensure that leaves are adjacent
-            if *right_index != left_index + 1 {
-                return Ok(Err(()));
-            }
-            // And that our target namespace is in between the leaves'
-            // namespaces
-            if namespace <= left_ns || namespace >= right_ns {
-                return Ok(Err(()));
-            }
-            // Verify the boundary proofs
-            if <InnerTree<E, H, T, N, Arity>>::verify(root, left_proof.index(), left_proof)?
-                .is_err()
-            {
-                return Ok(Err(()));
-            }
-            if <InnerTree<E, H, T, N, Arity>>::verify(root, right_proof.index(), right_proof)?
-                .is_err()
-            {
-                return Ok(Err(()));
-            }
-        }
-
-        Ok(Ok(()))
-    }
-
-    fn verify_presence_proof(
-        &self,
-        root: &NamespacedHash<T, N>,
-        namespace: N,
-    ) -> Result<VerificationResult, PrimitivesError> {
-        let mut last_idx: Option<u64> = None;
-        for (idx, proof) in self.proofs.iter().enumerate() {
-            let leaf_index = self.first_index + idx as u64;
-            if <InnerTree<E, H, T, N, Arity>>::verify(root, leaf_index, proof)?.is_err() {
-                return Ok(Err(()));
-            }
-            if proof
-                .elem()
-                .ok_or(PrimitivesError::InconsistentStructureError(
-                    "Missing namespace element".into(),
-                ))?
-                .get_namespace()
-                != namespace
-            {
-                return Ok(Err(()));
-            }
-            // Indices must be sequential, this checks that there are no gaps in the
-            // namespace
-            if let Some(prev_index) = last_idx {
-                if leaf_index != prev_index + 1 {
-                    return Ok(Err(()));
-                }
-                last_idx = Some(leaf_index);
-            }
-        }
-        // Verify that the proof contains the left boundary of the namespace
-        if self
-            .verify_left_namespace_boundary(root, namespace)
-            .is_err()
-        {
-            return Ok(Err(()));
-        }
-
-        // Verify that the proof contains the right boundary of the namespace
-        if self
-            .verify_right_namespace_boundary(root, namespace)
-            .is_err()
-        {
-            return Ok(Err(()));
-        }
-
-        Ok(Ok(()))
     }
 }
